add: drawNews.js
