+++ js/category.js
@@ -1,7 +1,7 @@
 import categoryData from "../json/category.json" assert { type: "json" };
 import news from "../json/news.json" assert { type: "json" };
-import news_article from "../json/news_article.json" assert { type: "json" };
 import { increment, totalTime } from "../utils/constants.js";
+import { drawNews } from "./drawNews.js";
 
 let progress = 0;
 let categoryCnt = []; //카테고리별 data개수
@@ -126,44 +126,6 @@ function progressFill(progressBar) {
   progressBar.style.width = `${progress}%`;
 }
 
-function drawNewsHeader() {
-  let PageNumberIndex = currentCategoryPageNumber - 1;
-  const news_header = document.querySelector(".news-header");
-  news_header.innerHTML = "";
-  let new_div = `<div class="news-header-div"><img class="news-thumbnail" src="${categoryNews[currentCategoryNumber][PageNumberIndex].thumbnail}"><span class="news-edit-time">${categoryNews[currentCategoryNumber][PageNumberIndex].editTime}</span><img class="subscribe-button" src="./img/subscribe_button.svg"></div>`;
-  news_header.innerHTML = new_div;
-}
-
-function drawNewsDiv() {
-  const PageNumberIndex = currentCategoryPageNumber - 1;
-  const listDiv = document.querySelector(".news-content");
-  listDiv.innerHTML = "";
-  let new_div = `<div class="news-main-left-div">${drawNewsImage(
-    PageNumberIndex
-  )}</div><div class="news-main-right-div">${drawNewsArticle(
-    PageNumberIndex
-  )}</div>`;
-  listDiv.innerHTML = new_div;
-}
-
-function drawNewsImage(PageNumberIndex) {
-  return `<img class="news-main-image" src="${categoryNews[currentCategoryNumber][PageNumberIndex].thumbnail}">${categoryNews[currentCategoryNumber][PageNumberIndex].title}`;
-}
-
-function drawNewsArticle(PageNumberIndex) {
-  let article_div = "";
-  for (let article_cnt = 0; article_cnt < 6; article_cnt++) {
-    article_div += `<div class="news-main-article">${news_article.article[article_cnt]}</div>`;
-  }
-  article_div += `<div class="news-main-article-press">${categoryNews[currentCategoryNumber][PageNumberIndex].press}언론사에서 직접 편집한 뉴스입니다.</div>`;
-  return article_div;
-}
-
-function drawNews() {
-  drawNewsDiv();
-  drawNewsHeader();
-}
-
 function clearCategoryNumber() {
   const categoryItem = document.querySelectorAll(".categoryItem");
   categoryDisplayClear(categoryItem);
@@ -237,8 +199,9 @@ function decreaseListPage() {
 export {
   drawInitCategory,
   progressInterval,
-  drawNewsHeader,
-  drawNewsDiv,
   increaseListPage,
   decreaseListPage,
+  currentCategoryNumber,
+  currentCategoryPageNumber,
+  categoryNews,
 };
